src/aws-waf-mcp-server/pyproject.toml
[project]
name = "mcp-waf"
version = "0.1.0"
description = "AWS WAF MCP server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.5",
    "mcp>=1.6.0",
    "moto>=5.1.4",
    "pytest-asyncio>=0.26.0",
]
